+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * added `input=datetime_immutable` to DateType, TimeType, DateTimeType
  * added `rounding_mode` option to MoneyType
+ * added translations for TimezoneType
 
 4.0.0
 -----